Split web_scanners.py to meet line limit
- Split web_scanners.py (302→195 lines) by extracting SqlmapAdapter
- Created sqlmap_adapter.py (107 lines) for SQL injection testing
- Updated web_scanners.py to import SqlmapAdapter for compatibility
- Progress: 4 modules still exceed 300-line limit, continuing refactoring
